changed logic of pseudo nov to prevent devisiopn through 0 v2

+++ engine/src/agents/mctsagent.h
@@ -84,8 +84,8 @@ public:
 
     
     float* factPlanes; //MR factPlanes stores the current score of all facts in plane representation
-    float timeStep; //MR save timestep for pseudocount novelty. It represents the call of the neural net
-    float featureProbabilities; //MR save product of all feature probabilities of timestep t
+    int timeStep; //MR save timestep for pseudocount novelty. It represents the call of the neural net
+    double featureProbabilities; //MR save product of all feature probabilities of timestep t
 
 public:
     MCTSAgent(NeuralNetAPI* netSingle,
